third_party/blink/web_tests/wpt_internal/display-lock/rendersubtree/inner-text.html

<!doctype HTML>
<html>
<meta charset="utf8">
<title>Display Locking: innerText</title>
<link rel="author" title="Vladimir Levin" href="mailto:vmpstr@chromium.org">
<link rel="help" href="https://github.com/WICG/display-locking">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>

This text should be visible.
<div id="container">
  This text should not be visible.
  <div id="inner">
    This text is also not visible.
  </div>
</div>

<script>
promise_test(async () => {
  const container = document.getElementById("container");
  await setInvisible(container);

  assert_equals(document.body.innerText, "This text should be visible.");
  assert_equals(document.getElementById("inner").innerText, "");
}, "innerText on locked element.");
</script>
</html>
